+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - utils.py: added convert_rhntw_to_rhcbci, the inverse of convert_rhcbci_to_rhntw, and a common function between the two: compute_R_rhntw_rhcbci to compute the rotation matrix they both use
 - private_src/: LBG1_LG6 environments and lbg1_private_utils
 
+### Changed:
+
+- plotters.py and runner.py: live telemetry viewer window will remain open for extended timeout (20sec), unless user manually closes, then it will forcefully close
+
 ## [v0.13.1] - 2025.10.21
 
 ### Added